Move distance transform to appropriate package
